+++ watchv2/Watch.h
@@ -22,26 +22,26 @@ class Watch {
     void update();
 
     private:
-    static const int MAX_SCREENS = 5;
+    static const byte MAX_SCREENS = 5;
 
-    static const int L_BUTTON_PIN = 7;
-    static const int R_BUTTON_PIN = 2;
+    static const byte L_BUTTON_PIN = 7;
+    static const byte R_BUTTON_PIN = 2;
 
     Screen *screens[MAX_SCREENS];
-    int current_screen;
-    int num_screens;
+    byte current_screen;
+    byte num_screens;
 
     WatchSystem watch_system;
 
     Button l_button;
     Button r_button;
-    int pot_pin;
+    byte pot_pin;
 
     void check_bluetooth();
     void add_screen(Screen *screen);
-    void switch_screen(int new_screen);
+    void switch_screen(byte new_screen);
     bool should_sleep(int *position);
 
     void update_screen();
-    void update_screen(bool button1_pressed, bool button2_pressed, int pot_pos);
+    void update_screen(bool button1_pressed, bool button2_pressed, short int pot_pos);
 };
